Create own sort by name util
